fix(ergo-ser): address CodeRabbit findings on #212 (trailing-bytes, k-width, wire caps)
Verified each finding against current code; fixed the still-valid ones:

- A: reject trailing bytes after ergoTree in tree_hash_from_bytes /
  template_hash_from_bytes (Scala consumes exactly `size`).
- B: template_hash returns Unparseable for an Unparsed body instead of
  hashing a partial tree.
- C: move the "Deserialize an ErgoTree from bytes" doc from the dangling
  spot above MAX_SUPPORTED_TREE_VERSION onto read_ergo_tree, which it
  describes.
- D: propagate embeddable_activated_version into the has_size read branch.
- F: replace debug-fragile assert!() caps with returned WriteError in
  write_coll and write_sigma_boolean (the latter made fallible; callers
  cascade). Matches Scala's u16 wire cap as a rejectable error, not a panic.
- G: SigmaBoolean::Cthreshold.k and UncheckedTree::Threshold.k widened
  u8 -> u16 to match Scala CTHRESHOLD(k: Int) / putUShort(th.k). The
  fiat-shamir tree still encodes k as a single byte (`*k as u8`), matching
  Scala UnprovenTree.toBytes `w.put(unchecked.k.toByte)` — verified against
  the reference so the consensus challenge bytes are unchanged.

Finding E (duplicate-val accept-invalid) is a deliberate, documented
tradeoff left as-is — see PR discussion; the naive "reject" fix would
reintroduce a reject-valid.

+++ ergo-sigma/src/evaluator/opcodes/sigma.rs
@@ -422,7 +422,7 @@ fn at_least_reduce(bound: i32, children: Vec<SigmaBoolean>) -> SigmaBoolean {
         return cand_normalized(sigmas);
     }
     SigmaBoolean::Cthreshold {
-        k: cur_bound as u8,
+        k: cur_bound as u16,
         children: sigmas,
     }
 }
